Fix two small bugs in move ordering
1. The relative history could overflow due to using 32-bit. Solution: switch to 64-bit.
2. The relative history score can exceed capture scores. Solution: force the move into the lower score list

+++ src/moveordering.hpp
@@ -21,8 +21,8 @@ namespace Arcanum
     class RelativeHistory
     {
         private:
-            uint32_t m_hhScores[2][64][64]; // History: Count the number of times the move did cause a Beta-cut
-            uint32_t m_bfScores[2][64][64]; // Butterfly: Count the number of times the move did not cause a Beta-cut
+            uint64_t m_hhScores[2][64][64]; // History: Count the number of times the move did cause a Beta-cut
+            uint64_t m_bfScores[2][64][64]; // Butterfly: Count the number of times the move did not cause a Beta-cut
         public:
             RelativeHistory();
             void addHistory(const Move& move, uint8_t depth, Color turn);
